+++ sdk/attestation/attestation/src/attestationClient.ts
@@ -8,9 +8,15 @@ import {
   AttestationSigner,
   AttestationToken,
   AttestationTokenValidationOptions,
-  AttestationResult,
-  AttestationData
+  AttestationResult
 } from "./models";
+
+import {
+  GeneratedAttestationResult,
+  InitTimeData,
+  KnownDataType,
+  RuntimeData
+} from "./generated/models";
 
 import { logger } from "./logger";
 import { createSpan } from "./tracing";
@@ -24,7 +30,8 @@ import { TypeDeserializer } from "./utils/typeDeserializer";
 import { TokenCredential } from "@azure/core-auth";
 import { CommonClientOptions, OperationOptions } from "@azure/core-client";
 import { bytesToString, stringToBytes } from "./utils/utf8";
-
+import { _attestationResultFromGenerated } from "./models/attestationResult";
+import { _attestationSignerFromGenerated } from "./models/attestationSigner";
 /**
  * Attestation Client Construction Options.
  */
@@ -56,14 +63,26 @@ export interface AttestationClientOperationOptions extends OperationOptions {
  */
 export interface AttestOpenEnclaveOptions extends AttestationClientOperationOptions {
   /**
-   *initTimeData : AttestationData - data provided at the time the enclave was initialized.
+   *initTimeData : data provided at the time the enclave was initialized, to be interpreted as binary data.
    *
    */
-  initTimeData?: AttestationData;
+  initTimeData?: Uint8Array;
+
   /**
-   * runTimeData  - data provided at the time the SGX quote being attested was created.
+   * inittimeJson : data provided at the time the enclave was initialized, to be interpreted as JSON data.
    */
-  runTimeData?: AttestationData;
+  initTimeJson?: Uint8Array;
+
+  /**
+   * runTimeData  - data provided at the time the OpenEnclave report being attested was created to be interpreted as binary data.
+   */
+  runTimeData?: Uint8Array;
+
+  /**
+   * runTimeJson  - data provided at the time the OpenEnclave report being attested was created to be interpreted as JSON data.
+   */
+  runTimeJson?: Uint8Array;
+
   /**
    * draftPolicyForAttestation - If specified, the attestation policy to be used during the attestation request.
    */
@@ -79,15 +98,25 @@ export interface AttestOpenEnclaveOptions extends AttestationClientOperationOpti
  */
 export interface AttestSgxEnclaveOptions extends AttestationClientOperationOptions {
   /**
-   *initTimeData : AttestationData - data provided at the time the enclave was initialized.
+   *initTimeData : data provided at the time the enclave was initialized, to be interpreted as binary data.
    *
    */
-  initTimeData?: AttestationData;
+  initTimeData?: Uint8Array;
 
   /**
-   * runTimeData - data provided at the time the SGX quote being attested was created.
+   * inittimeJson : data provided at the time the enclave was initialized, to be interpreted as JSON data.
    */
-  runTimeData?: AttestationData;
+  initTimeJson?: Uint8Array;
+
+  /**
+   * runTimeData  - data provided at the time the OpenEnclave report being attested was created to be interpreted as binary data.
+   */
+  runTimeData?: Uint8Array;
+
+  /**
+   * runTimeJson  - data provided at the time the OpenEnclave report being attested was created to be interpreted as JSON data.
+   */
+  runTimeJson?: Uint8Array;
 
   /**
    * draftPolicyForAttestation - If specified, the attestation policy to be used during the attestation request.
@@ -172,22 +201,51 @@ export class AttestationClient {
     options: AttestOpenEnclaveOptions = {}
   ): Promise<AttestationResponse<AttestationResult>> {
     const { span, updatedOptions } = createSpan("AttestationClient-attestOpenEnclave", options);
+
     try {
+      if (options.initTimeData !== undefined && options.initTimeJson !== undefined) {
+        throw new TypeError("Cannot provide both initTimeData and initTimeJson.");
+      }
+
+      if (options.runTimeData !== undefined && options.runTimeJson !== undefined) {
+        throw new TypeError("Cannot provide both runTimeData and runTimeJson.");
+      }
+
+      if (options.initTimeJson !== undefined) {
+        try {
+          JSON.parse(options.initTimeJson.toString());
+        } catch (e) {
+          throw new TypeError("initTimeJson value cannot be parsed as JSON " + e.message);
+        }
+      }
+
+      if (options.runTimeJson !== undefined) {
+        try {
+          JSON.parse(bytesToString(options.runTimeJson));
+        } catch (e) {
+          throw new TypeError("runTimeJson value cannot be parsed as JSON " + e.message);
+        }
+      }
+
+      const initData = options.initTimeData ?? options.initTimeJson;
+      const initTimeData: InitTimeData | undefined = initData
+        ? { data: initData, dataType: options.initTimeJson !== undefined ? KnownDataType.Json : KnownDataType.Binary
+        }
+        : undefined;
+
+      const runData = options.runTimeData ?? options.runTimeJson;
+      const runTimeData: RuntimeData | undefined = runData
+        ? {
+            data: runData,
+            dataType: options.runTimeJson !== undefined ? KnownDataType.Json : KnownDataType.Binary
+        }
+        : undefined;
+
       const attestationResponse = await this._client.attestation.attestOpenEnclave(
         {
           report: report,
-          initTimeData: options.initTimeData
-            ? {
-                data: options.initTimeData.data,
-                dataType: options.initTimeData.isJson ? "JSON" : "Binary"
-              }
-            : undefined,
-          runtimeData: options.runTimeData
-            ? {
-                data: options.runTimeData.data,
-                dataType: options.runTimeData.isJson ? "JSON" : "Binary"
-              }
-            : undefined,
+          initTimeData: initTimeData,
+          runtimeData: runTimeData,
           draftPolicyForAttestation: options.draftPolicyForAttestation ?? undefined
         },
         updatedOptions
@@ -195,20 +253,23 @@ export class AttestationClient {
 
       const token = new AttestationToken(attestationResponse.token);
       token.validateToken(
-        await this._signingKeys,
+        await this._signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
 
       const attestationResult = TypeDeserializer.deserialize(
         token.getBody(),
         {
-          AttestationResult: Mappers.AttestationResult,
+          GeneratedAttestationResult: Mappers.GeneratedAttestationResult,
           JsonWebKey: Mappers.JsonWebKey
         },
-        "AttestationResult"
-      ) as AttestationResult;
+        "GeneratedAttestationResult"
+      ) as GeneratedAttestationResult;
 
-      return new AttestationResponse<AttestationResult>(token, attestationResult);
+      return new AttestationResponse<AttestationResult>(
+        token,
+        _attestationResultFromGenerated(attestationResult)
+      );
     } catch (e) {
       span.setStatus({ code: SpanStatusCode.ERROR, message: e.message });
       throw e;
@@ -230,21 +291,49 @@ export class AttestationClient {
   ): Promise<AttestationResponse<AttestationResult>> {
     const { span, updatedOptions } = createSpan("AttestationClient-attestSgxEnclave", options);
     try {
+      if (options.initTimeData !== undefined && options.initTimeJson !== undefined) {
+        throw new TypeError("Cannot provide both initTimeData and initTimeJson.");
+      }
+
+      if (options.runTimeData !== undefined && options.runTimeJson !== undefined) {
+        throw new TypeError("Cannot provide both runTimeData and runTimeJson.");
+      }
+
+      if (options.initTimeJson !== undefined) {
+        try {
+          JSON.parse(bytesToString(options.initTimeJson));
+        } catch (e) {
+          throw new TypeError("initTimeJson value cannot be parsed as JSON " + e.message);
+        }
+      }
+
+      if (options.runTimeJson !== undefined) {
+        try {
+          JSON.parse(bytesToString(options.runTimeJson));
+        } catch (e) {
+          throw new TypeError("runTimeJson value cannot be parsed as JSON " + e.message);
+        }
+      }
+
+      const initData = options.initTimeData ?? options.initTimeJson;
+      const initTimeData: InitTimeData | undefined = initData
+        ? { data: initData, dataType: options.initTimeJson !== undefined ? KnownDataType.Json : KnownDataType.Binary
+        }
+        : undefined;
+
+      const runData = options.runTimeData ?? options.runTimeJson;
+      const runTimeData: RuntimeData | undefined = runData
+        ? {
+            data: runData,
+            dataType: options.runTimeJson !== undefined ? KnownDataType.Json : KnownDataType.Binary
+        }
+        : undefined;
+
       const attestationResponse = await this._client.attestation.attestSgxEnclave(
         {
           quote: quote,
-          initTimeData: options.initTimeData
-            ? {
-                data: options.initTimeData.data,
-                dataType: options.initTimeData.isJson ? "JSON" : "Binary"
-              }
-            : undefined,
-          runtimeData: options.runTimeData
-            ? {
-                data: options.runTimeData.data,
-                dataType: options.runTimeData.isJson ? "JSON" : "Binary"
-              }
-            : undefined,
+          initTimeData: initTimeData,
+          runtimeData: runTimeData,
           draftPolicyForAttestation: options.draftPolicyForAttestation ?? undefined
         },
         updatedOptions
@@ -252,20 +341,23 @@ export class AttestationClient {
 
       const token = new AttestationToken(attestationResponse.token);
       token.validateToken(
-        await this._signingKeys,
+        await this._signingKeys(),
         options.validationOptions ?? this._validationOptions
       );
 
       const attestationResult = TypeDeserializer.deserialize(
         token.getBody(),
         {
-          AttestationResult: Mappers.AttestationResult,
+          GeneratedAttestationResult: Mappers.GeneratedAttestationResult,
           JsonWebKey: Mappers.JsonWebKey
         },
-        "AttestationResult"
-      ) as AttestationResult;
+        "GeneratedAttestationResult"
+      ) as GeneratedAttestationResult;
 
-      return new AttestationResponse<AttestationResult>(token, attestationResult);
+      return new AttestationResponse<AttestationResult>(
+        token,
+        _attestationResultFromGenerated(attestationResult)
+      );
     } catch (e) {
       span.setStatus({ code: SpanStatusCode.ERROR, message: e.message });
       throw e;
@@ -331,7 +423,7 @@ export class AttestationClient {
       const signingCertificates = await this._client.signingCertificates.get(updatedOptions);
       const signers: AttestationSigner[] = new Array();
       signingCertificates.keys?.forEach((element) => {
-        signers.push(new AttestationSigner(element));
+        signers.push(_attestationSignerFromGenerated(element));
       });
       return signers;
     } catch (e) {
@@ -361,13 +453,18 @@ export class AttestationClient {
 
   private _client: GeneratedClient;
   private _validationOptions?: AttestationTokenValidationOptions;
-  private _signers?: Promise<AttestationSigner[]>;
+  private _signers?: AttestationSigner[];
 
-  private get _signingKeys(): Promise<AttestationSigner[]> {
+  private async _signingKeys(): Promise<AttestationSigner[]> {
     if (this._signers !== undefined) {
-      return Promise.resolve(this._signers);
+      return this._signers;
     }
-    this._signers = this.getAttestationSigners();
-    return Promise.resolve(this._signers);
+    const jwks = await this._client.signingCertificates.get();
+    const signers: AttestationSigner[] = new Array();
+    jwks.keys?.forEach((element) => {
+      signers.push(_attestationSignerFromGenerated(element));
+    });
+    this._signers = signers;
+    return this._signers;
   }
 }