+++ sdk/attestation/attestation/src/models/policyResult.ts
@@ -13,18 +13,24 @@ import * as Mappers from "../generated/models/mappers";
 import { PolicyResult as GeneratedPolicyResult } from "../generated/models";
 
 import { TypeDeserializer } from "../utils/typeDeserializer";
-import { AttestationSigner } from "./attestationSigner";
+import { AttestationSigner, _attestationSignerFromGenerated } from "./attestationSigner";
 
 /**
  * The result of a policy certificate modification
  */
 export class PolicyResult {
-  private constructor(result: GeneratedPolicyResult) {
+  /**
+   *
+   * @param result Generated PolicyResult from service.
+   *
+   * @hideconstructor
+   */
+  constructor(result: any) {
     this.policyResolution = result.policyResolution;
     this.policyTokenHash = result.policyTokenHash;
     this.policy = result.policy;
     this.policySigner = result.policySigner
-      ? new AttestationSigner(result.policySigner)
+      ? _attestationSignerFromGenerated(result.policySigner)
       : undefined;
   }
 
@@ -44,20 +50,20 @@ export class PolicyResult {
    * A JSON Web Token containing a StoredAttestationPolicy object with the attestation policy
    */
   policy?: string;
+}
 
-  /**
-   * Create a new instance of a {@link PolicyResult} from a raw JSON object.
-   * @param rawJson - JSON Policy Result object returned from the Attestation service.
-   * @returns a newly created {@link PolicyResult} object whose contents reflect the JSON received from
-   *  the attestation service.
-   */
-  static create(rawJson: unknown): PolicyResult {
-    return new PolicyResult(
-      TypeDeserializer.deserialize(
-        rawJson,
-        { PolicyResult: Mappers.PolicyResult, JsonWebKey: Mappers.JsonWebKey },
-        "PolicyResult"
-      ) as GeneratedPolicyResult
-    );
-  }
+/**
+ * Create a new instance of a {@link PolicyResult} from a raw JSON object.
+ * @param rawJson - JSON Policy Result object returned from the Attestation service.
+ * @returns a newly created {@link PolicyResult} object whose contents reflect the JSON received from
+ *  the attestation service.
+ */
+export function _policyResultFromGenerated(rawJson: unknown): PolicyResult {
+  return new PolicyResult(
+    TypeDeserializer.deserialize(
+      rawJson,
+      { PolicyResult: Mappers.PolicyResult, JsonWebKey: Mappers.JsonWebKey },
+      "PolicyResult"
+    ) as GeneratedPolicyResult
+  );
 }